refactor(eslint-rule): rename and make check more flexible
