--- packages/eslint-plugin-await-upsert/src/index.ts
@@ -1,7 +1,0 @@
-import { TSESLint } from "@typescript-eslint/utils";
-
-import rule from "./rule";
-
-export const rules = {
-  "await-upsert": rule,
-} satisfies Record<string, TSESLint.RuleModule<string, Array<unknown>>>;
