+++ packages/eslint-plugin-indexing/.gitignore
