+++ packages/eslint-plugin-indexing/src/index.ts
@@ -1,0 +1,5 @@
+import rule from "./rule";
+
+export const rules = {
+  "await-upsert": rule,
+};
