+++ packages/eslint-plugin-indexing/tsconfig.json
